requirements.txt
requests
pydantic
click
python-dotenv
pyairtable
rapidfuzz
beautifulsoup4
lxml
tenacity
rich
pyyaml
